requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
flake8
black
bcrypt==4.0.1
passlib
python-jose[cryptography]


